Fix compile error due to bit shift in constant
The 1 defaults to int, which is signed, so shifting it by 31 bits is too
much. This probably was a warning before, and then it would probably
work fine because the result is cast to an unsigned int, but with recent
gcc/arduino versions this produces a compilation error.

By casting to unsigned first, this error is avoided.

+++ src/SleepHandler.h
@@ -50,7 +50,7 @@ class SleepHandler {
     }
 
     // Maximum time that a single sleep can last (±9.5 hours)
-    static const uint32_t MAX_SLEEP_TICKS = (1 << 31);
+    static const uint32_t MAX_SLEEP_TICKS = ((uint32_t)1 << 31);
 
     // Returns the number of ticks since startup
     static uint32_t ticks() {
